Cache images using MD5 hash instead of random UUID
Use an MD5 hash of the Bitmap contents instead of a random UUID for each
image, so that multiple images will not be stored on disk for the same
resource. If the same content is use for an additional image field of
DataCard or DataCardImage, the existing cached image will be reused.

+++ sdk/src/org/cyanogenmod/launcher/home/api/provider/CmHomeContentProvider.java
@@ -20,11 +20,16 @@ import android.util.Log;
 import org.cyanogenmod.launcher.home.api.cards.DataCardImage;
 import org.cyanogenmod.launcher.home.api.db.CmHomeDatabaseHelper;
 
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.OutputStream;
+import java.math.BigInteger;
 import java.net.URI;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
@@ -377,16 +382,79 @@ public class CmHomeContentProvider extends ContentProvider {
         }
     }
 
-    public static Uri storeBitmapInCache(Bitmap bitmap, Context context) {
-        String filename = UUID.randomUUID().toString() + ".png";
-        FileOutputStream outputStream = null;
+    /**
+     * Generates a String representing the MD5 hash of the input byte array.
+     * @param bytes A byte array
+     * @return A String representation of the MD5 hash of the input byte array
+     */
+    private static String hashBytesMD5(byte[] bytes) {
+        // Compute the hash
         try {
+            MessageDigest md = MessageDigest.getInstance("MD5");
+            md.update(bytes, 0, bytes.length);
+            String hash = new BigInteger(1, md.digest()).toString(16);
+            return hash;
+        } catch (NoSuchAlgorithmException e) {
+            Log.w(TAG, "Unable to compute MD5 hash of byte array.");
+        }
+
+        // No hash computed
+        return null;
+    }
+
+    /**
+     * Check if a cached Bitmap file exists in {@link org.cyanogenmod.launcher.home.api.provider
+     * .CmHomeContentProvider.IMAGE_FILE_CACHE_DIR} with the given filename.
+     * @param filename The filename to check for
+     * @param context The context with access to the directory that would contain this file.
+     * @return True if the file exists.
+     */
+    private static boolean bitmapCacheFileExists(String filename, Context context) {
+        // Create a file in the cache subdirectory
+        File imageDir = new File(context.getFilesDir(), IMAGE_FILE_CACHE_DIR);
+        File imageFile = new File(imageDir, filename);
+
+        return imageFile.exists();
+    }
+
+    /**
+     * Stores the given bitmap in internal storage in {@link org.cyanogenmod.launcher
+     * .home.api.provider.CmHomeContentProvider.IMAGE_FILE_CACHE_DIR} using an MD5 sum of the
+     * bitmap content as the filename, if the cache does not exist already.
+     * @param bitmap The Bitmap to store in the cache
+     * @param context A Context of the application that will share this image in this
+     *                ContentProvider.
+     * @return A Uri pointing to the newly stored image file in the cache, or the existing image,
+     * if one is found.
+     */
+    public static Uri storeBitmapInCache(Bitmap bitmap, Context context) {
+        FileOutputStream outputStream = null;
+        ByteArrayOutputStream byteArrayOutputStream = null;
+        try {
+            // Get the bytes containing the image data
+            byteArrayOutputStream = new ByteArrayOutputStream();
+            bitmap.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOutputStream);
+            byte[] bitmapBytes = byteArrayOutputStream.toByteArray();
+
+            String hash = hashBytesMD5(bitmapBytes);
+            // Can't continue without a hash
+            if (hash == null) return null;
+
+            String filename = hash + ".png";
+
+            // If the cache already exists, just return the URI to the cache file
+            if (bitmapCacheFileExists(filename, context)) {
+                return Uri.withAppendedPath(CmHomeContract.ImageFile.CONTENT_URI,
+                                            filename);
+            }
+
+            // Write the bytes to a file using the hash as the filename
             // Create a file in the cache subdirectory
             File imageDir = new File(context.getFilesDir(), IMAGE_FILE_CACHE_DIR);
             imageDir.mkdirs();
             File imageFile = new File(imageDir, filename);
             outputStream = new FileOutputStream(imageFile);
-            bitmap.compress(Bitmap.CompressFormat.PNG, 100, outputStream);
+            outputStream.write(bitmapBytes);
 
             Uri imageUri = Uri.withAppendedPath(CmHomeContract.ImageFile.CONTENT_URI,
                                                 filename);
@@ -394,15 +462,20 @@ public class CmHomeContentProvider extends ContentProvider {
             // Set the image URI, which will actually be stored in the database.
             return imageUri;
         } catch (FileNotFoundException e) {
-            Log.e(TAG, "Unable to save bitmap to temporary file.");
+            Log.e(TAG, "Unable to save bitmap to temporary file. Could not open file.");
+        } catch (IOException e) {
+            Log.e(TAG, "Unable to save bitmap to temporary file, IOException occurred.");
         } finally {
-            if (outputStream != null) {
                 try {
-                    outputStream.close();
+                    if (outputStream != null) {
+                        outputStream.close();
+                    }
+                    if (byteArrayOutputStream != null) {
+                        byteArrayOutputStream.close();
+                    }
                 } catch (IOException e) {
-                    Log.e(TAG, "Unable to save bitmap to temporary file.");
+                    Log.e(TAG, "Unable to save bitmap to temporary file, IOException occurred.");
                 }
-            }
         }
         // Failure, no URI available
         return null;
